+++ Tests/FeatherWidgets.TestUI.Arrangements/Packaging/StaticContent/Libraries/ImportEditedVideosStructure.cs
@@ -49,10 +49,14 @@ namespace FeatherWidgets.TestUI.Arrangements
         /// </summary>
         [ServerTearDown]
         public void ClearUp()
-        {            
-            ServerOperations.Videos().DeleteLibrary(AlbumName);
+        {
+            AuthenticationHelper.AuthenticateUser(AdminUserName, AdminPass, true);
+            ServerOperations.Videos().DeleteLibrary(AlbumName, false);
             ServerOperations.Pages().DeleteAllPages();
-            ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
+            if (System.IO.Directory.Exists(this.tempFolderPath))
+                ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
             ServerOperations.Packaging().DeleteAllPackagesFromDB();
 
             for (int i = 0; i < this.widgetTemplatesNames.Length; i++)
@@ -89,6 +93,8 @@ namespace FeatherWidgets.TestUI.Arrangements
             ServerOperations.Taxonomies().DeleteFlatTaxonomy(flatClassificationVideo);
         }
 
+        private const string AdminUserName = "admin";
+        private const string AdminPass = "admin@2";
         private const string InstallationPath = @"App_Data\Sitefinity";
         private const string PackageResource = "FeatherWidgets.TestUtilities.Data.Packaging.Structure.LibrariesStructure.zip";
         private const string PackageResourceEdited = "FeatherWidgets.TestUtilities.Data.Packaging.Structure.LibrariesEdited.zip";
