+++ Tests/FeatherWidgets.TestUI.Arrangements/Packaging/StaticContent/Pages/ExportEditedPages.cs
@@ -2,6 +2,7 @@
 using System.Web;
 using System.Web.Hosting;
 using FeatherWidgets.TestUtilities.CommonOperations;
+using Telerik.Sitefinity.Restriction;
 using Telerik.Sitefinity.TestArrangementService.Attributes;
 using Telerik.Sitefinity.TestUI.Arrangements.Framework;
 using Telerik.Sitefinity.TestUtilities.CommonOperations;
@@ -44,21 +45,30 @@ namespace FeatherWidgets.TestUI.Arrangements
         [ServerTearDown]
         public void ClearUp()
         {
+            AuthenticationHelper.AuthenticateUser(AdminUserName, AdminPass, true);
             ServerOperations.Pages().DeleteAllPages();
-            ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
+            if (System.IO.Directory.Exists(this.tempFolderPath))
+                ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
             ServerOperations.Packaging().DeleteAllPackagesFromDB();
 
-            for (int i = 0; i < ServerOperations.CustomFieldsNames().FieldNamesWithoutClassificationsEdited.Length; i++)
+            using (new UnrestrictedModeRegion())
             {
-                ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, ServerOperations.CustomFieldsNames().FieldNamesWithoutClassificationsEdited[i]);
-                ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, flatClassification);
-                ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, hierarchicalClassification);
+                for (int i = 0; i < ServerOperations.CustomFieldsNames().FieldNamesWithoutClassificationsEdited.Length; i++)
+                {
+                    ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, ServerOperations.CustomFieldsNames().FieldNamesWithoutClassificationsEdited[i]);
+                    ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, flatClassification);
+                    ServerOperations.CustomFields().RemoveCustomFieldsFromContent(PagesType, hierarchicalClassification);
+                }
             }
 
             ServerOperations.Taxonomies().DeleteHierarchicalTaxonomy(hierarchicalClassification);
             ServerOperations.Taxonomies().DeleteFlatTaxonomy(flatClassification);
         }
 
+        private const string AdminUserName = "admin";
+        private const string AdminPass = "admin@2";
         private const string InstallationPath = @"App_Data\Sitefinity";
         private const string PackageResource = "FeatherWidgets.TestUtilities.Data.Packaging.Structure.PagesStructure.zip";
         private const string PackageResourceEdited = "FeatherWidgets.TestUtilities.Data.Packaging.Structure.PagesEdited.zip";
